Determine and provide latitude/longitude on map
This is the first working draft of a new feature allowing the user to
determine the latitude and longitude by specifying the Cartesian
coordinates of the projection.

For example displaying the user the current location on mouse move or
handling the location for further tasks after a click (e.g. marker).

Changes / TODOs:
  * Implemented the inverse function of the "Miller" projection.
  * TODO: The inverse functions of the 3 other projections are not yet
    implemented (merc, aea, lcc).
  * Implemented access function pointToLatLng(x, y).
  * Instead of the worlds circumference	the radius is used in
    the algorithm.
  * Changed the worlds default radius.
  * Added mousemove-events for regions and markers.
  * Added testpage latlng.html as a demonstration.
  * Some features were not (thoroughly) tested. Especially the insets
    and boundingbox offsets should be tested.