Smoother zooming out when resizing the map
When resizing the map, the viewport is currently moving to one size. That
means that if you have the map centered on some country, resize the map to
make it smaller, than resize it to the original size and you are centered
on a different country.

With this patch, it tries to keep the viewport at the same point, so that
resizing the map has little effect on what you see on the map.