Suggested fix for "jquery.jvectormap.js:1929 Uncaught TypeError: Cannot read property 'element' of undefined"
This error occurs when the data you hand over to the map as selectedRegions contains country short codes which the current map does not have. In our application e.g. our webservice always delivers data from all countries of the world, but we're using a map of Europe only. Diffing our data with the available countries in our chosen map would be painful.